refactor: move create server and client funcs into own files
